+++ src/main/java/undc/objects/ObstacleItem.java
@@ -5,7 +5,7 @@ import javafx.scene.image.Image;
 public class ObstacleItem extends Obstacle {
     private Item item;
 
-    public ObstacleItem(Image sprite, double x, double y, double w, double h, ObstacleType type) {
+    public ObstacleItem(Image sprite, double x, double y, int w, int h, ObstacleType type) {
         super(sprite, x, y, w, h, type);
     }
